RDKB-42081: Porting OneWifi cleanup changes to stable2 branch
Reason for change: Disable dml, stats compilation, remove ccsp-common dependencies
Test Procedure: Verify XB7 and XE2 functionality first to verify OneWifi is not broken

Risks: High


RDKB-44716: Get rid of source/platform folder

Reason for change: OneWifi cleanup
Test Procedure: Verify XB7 and XE2 functionality first to verify OneWifi is not broken

Risks: High

+++ lib/const/const.h
@@ -100,7 +100,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 // Return number of elements in array
 #ifndef ARRAY_SIZE
-#define ARRAY_SIZE(x)       (sizeof(x) / sizeof(x[0]))
+#define ARRAY_SIZE(x)       ((unsigned int)(sizeof(x) / sizeof(x[0])))
 #endif /* ARRAY_SIZE */
 
 // Same as ARRAY_SIZE, except returned signed value
